TMDM-13735 Remove power mock from unit test.
